+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change log for SharePointDsc
 
 ## Unreleased
-* Created DSCResource for Access Services 2010
 
+* Created DSCResource for Access Services 2010
 * Updated documentation in regards to guidance on installing binaries from
   network locations instead of locally
 * New resources: SPFarmPropertyBag